First pass on ideas
checkboxes are spiteful elements.
Need to figure out idea for setting queryParams for initialization when checkboxes are default checked in the template. Current solution only allows for 1 checkbox to be default checked at a time...

Want to look at if current reactive forms/CVA functionality works on component libraries that use CVAs.